wip delete account bug

+++ client/src/components/user-settings/DeleteAccount.jsx
@@ -19,9 +19,17 @@ export default function DeleteAccount() {
   const { backend } = useBackendContext();
   const { currentUser } = useAuthContext();
   const [input, setInput] = useState("");
-  const handleDelete = () => {
-    backend.delete(`/users/${currentUser.uid}`);
+
+  const handleDelete = async () => {
+    try {
+      if (input === "DELETE") {
+        await backend.delete(`/users/${currentUser.uid}`);
+      }
+    } catch (e) {
+      alert("Incorrect.");
+    }
   };
+
   return (
     <Grid>
       <FormControl>
@@ -30,13 +38,22 @@ export default function DeleteAccount() {
           <Text>
             Are you sure? Deleting your account is permament and will remove all
             your information from the database. This action{" "}
-            <Text as={"span"} fontWeight={"bold"}>cannot</Text> be undone.
+            <Text
+              as={"span"}
+              fontWeight={"bold"}
+            >
+              cannot
+            </Text>{" "}
+            be undone.
           </Text>
           <Text>To confirm this, type "DELETE"</Text>
-          <Input value={input} />
+          <Input
+            value={input}
+            onChange={(e) => setInput(e.target.value)}
+          />
         </Grid>
+        <Button onClick={handleDelete}>Delete Account</Button>
       </FormControl>
-      <Button onClick={handleDelete}>Delete Account</Button>
     </Grid>
   );
 }
